+++ src/cp/constraints/same_sequence_on_each_machine.py
